Extract about-modal region into sections/about-modal.js
Second of P-08's six region extractions. 20 LOC — the smallest of the batch, no cross-region references, single document-level click listener wrapped in initAboutModalSideEffects().
